# fix: speed up "fast" linters