Currently, there are 2 problems:

1. `typecheck` uses `LoadModeTypesInfo` and `WithLoadForGoAnalysis`.
   So the load mode of the linter config for the `metalinter` is always `LoadModeTypesInfo` and never `LoadModeSyntax`.
2. the `metalinter` has a hardcoded call to `WithLoadForGoAnalysis`, then the types are always loaded.  

As a consequence, the "fast" linters (like `lll`, `gofmt`, `misspell`, etc.) are slower than expected when either they are running alone or only "fast" linters are run.

This PR changes the config linter load mode of `typecheck` to `LoadModeNone` and removes `WithLoadForGoAnalysis` for this fake linter.
The `metalinter` goanalysis load mode is computed based on linters.

`typecheck` doesn't need an explicit load mode because it is never run alone, then at the end, the load mode will depend on the load mode of the real linters.
Same thing for the goanalysis load mode.

FYI, there are 2 load modes:
- the config linter load mode (`LoadModeTypesInfo`, `LoadModeSyntax`, etc.)
- the goanalysis load mode (`WithLoadForGoAnalysis`, `WithLoadFiles`)

They are used for different purposes.

---

Based on local tests (verbose mode), the performances are improved (~50%).

<details>

### On golangci-lint code

Current version:
- `golangci-lint cache clean; golangci-lint run --enable-only misspell --enable-only lll  -v`
-  Execution took between 1.4s and 950ms

With the PR:
- `golangci-lint cache clean; ./golangci-lint run --enable-only misspell --enable-only lll  -v`
-  Execution took between 900ms and 800ms


### On Traefik code

Current version:
- `golangci-lint cache clean; golangci-lint run --enable-only gofmt --enable-only misspell -v`
-  Execution took between 1.3s and 1.2ms

Current version:
- `golangci-lint cache clean; ./golangci-lint run --enable-only gofmt --enable-only misspell -v`
-  Execution took between 650ms and 500ms

</details>

Fixes #2914
Fixes #4639